fix user tracking tests (#5497)
